Split reconnect test seams
