+++ Packages/iOS/CmuxMobileShell/Sources/CmuxMobileShell/MobileShellComposite+ReconnectPresentation.swift
@@ -1,0 +1,11 @@
+import Foundation
+
+extension MobileShellComposite {
+    public var shouldPreserveWorkspaceShellDuringReconnect: Bool {
+        connectionState != .connected
+            && hasCachedRemoteWorkspaceSnapshot
+            && !connectionRequiresReauth
+            && (isRecoveringConnection || isReconnectingStoredMac)
+            && workspaces.contains { !$0.terminals.isEmpty }
+    }
+}
