+++ Packages/iOS/CmuxMobileShellUI/Sources/CmuxMobileShellUI/CMUXMobileRootView.swift
@@ -70,26 +70,6 @@ struct CMUXMobileRootView: View {
         #endif
     }
 
-    @ViewBuilder private var terminalLayoutPreview: some View {
-        #if os(iOS) && DEBUG
-        TerminalLayoutPreviewView()
-        #else
-        EmptyView()
-        #endif
-    }
-
-    // `WorkspaceListLayoutPreviewView` is `#if DEBUG`-only (a simulator
-    // screenshot fixture), so referencing it directly in `rootContent` breaks the
-    // Release archive ("cannot find ... in scope"). Gate the reference here, the
-    // same way `terminalLayoutPreview` does, so Release compiles to `EmptyView`.
-    @ViewBuilder private var workspaceListLayoutPreview: some View {
-        #if os(iOS) && DEBUG
-        WorkspaceListLayoutPreviewView()
-        #else
-        EmptyView()
-        #endif
-    }
-
     var body: some View {
         rootContent
         .sheet(isPresented: addDeviceSheetBinding) {
